// src/enums/SetType.ts
export enum SetType {
    SET_HEALING = "passerby",
    SET_MUSKEETER = "muskeeter",
    SET_DEF = "knight",
    SET_ICE = "hunter",
    SET_PHSYICAL = "champion",
    SET_DMGTAKEN = "guard",
    SET_FIRE = "firesmith",
    SET_QUANTUM = "genius",
    SET_LIGHTNING = "band",
    SET_WIND = "eagle",
    SET_THIEF = "thief",
    SET_IMAGINARY = "Wastelander",
    SET_HP = "longevous",
    SET_MESSENGER = "messenger",
    SET_FOLLOWUP = "ashblazing",
    SET_DOT = "prisoner",
    SET_DEBUFF = "pioneer",
    SET_BREAKEFFECT = "watchmaker",
    PLANAR_STATION = "spaceStation",
    PLANAR_HP = "fleet",
    PLANAR_EHR = "pancosmic",
    PLANAR_DEF = "belobog",
    PLANAR_CD = "celestialDifferentiator",
    PLANAR_SALSOTTO = "salsotto",
    PLANAR_TALIA = "talia",
    PLANAR_VONWACQ = "vonwacq",
    PLANAR_RUTILANT = "rutilant",
    PLANAR_KEEL = "keel",
    PLANAR_GLAMOTH = "glamoth",
    PLANAR_SAMETYPE = "penacony",
    PLANAR_PUREFICTION = "sigonia",
    PLANAR_SAMEPATH = "izumo"
}